Perform phylogenetic pair analysis

Perform phylogenetic pair analysis

PP(z, tree = NULL, dists = NULL, pp = NULL, seed = NA, zName = "z",
  treeName = "tree", distsName = "dists", tauQuantiles = c(V = 0.05, D =
  0.1, O = 0.125, qu = 0.2, Q = 0.25, M = 0.5, A = 1), bootstraps = 0,
  exclude = data.table(name = c("all"), scopeAll = c("FALSE"), scopeTau =
  c("FALSE"), key = "name"), verbose = FALSE, ...)

Arguments

exclude
a data.table specifying selection filters to be applied after the extraction of phylogenetic pairs. Every row specifies one such filter in the form of R-expressions to be evaluated within the data.table of extracted phylogenetic pairs (see exctractPP for the description of columns available in one such data.table). For each row in the pp-table, for which an expression evaluates to TRUE, the corresponding phylogenetic pair rows (the row itself and its partner row) get removed from the ANOVA analysis. The exclude data.table has three character vector columns as follows: name - meaningful name of the filter used as an index (key) in the data.table, scopeAll - R-expression evaluated before grouping by quantiles of tau. scopeTau - R-expression evaluated within each tau-quantile group. Note that the filters are applied on the pp data.table or after a call to extractPP, so they cannot affect the formation of phylogenetic pairs. Note also that if a member of a phylogenetic pair gets excluded its pair-partner is also removed even if the filter expression does not evaluate to TRUE for it. For example the filter filters=data.table(name=c('all', 'no outliers in deltaz'), scopeAll=('FALSE'), scopeTau=c('FALSE', 'deltaz>{q=quantile(deltaz); q[4]+1.5*(q[4]-q[2])}'), key='name') would result in a PP analysis on all phylogenetic pairs, and a PP analysis on the phylogenetic pairs which's phenotypic distance deltaz doesn't exceed .75 filtering is done.

...
currently not used

Value

a data.table with the estimated statistics for each tauQuantile (and filter).