Broad-sense heritability of a pathogen trait

Broad-sense heritability of a pathogen trait

R2adj(epidemic = NULL, data = NULL, GEValues = NULL, sampledOnly = TRUE,
  activeOnly = FALSE, tMin = 0, tMax = Inf, firstN = Inf, lastN = Inf)

Arguments

epidemic
a list of objects returned from simulateEpidemic (see toyepidemic package)

data
NULL or a data.table such as the element gen in an epidemic list

GEValues
genotype-environment trait values for calculating the z-values; if NULL it is assumed that the z-values are already in data, or epidemic$gen.

sampledOnly, activeOnly, tMin, tMax, firstN, lastN
parameters passed to extractPop

Value

numeric indicating the estimated heritability for infected individuals in the population.