Regression slope of recipient on donor values in an epidemic simulated with the toyepidemic R-package.

Regression slope of recipient on donor values in an epidemic simulated with the toyepidemic R-package.

b(epidemic = NULL, data = NULL, GEValues, sampledOnly = TRUE,
  activeOnly = FALSE, tMin = 0, tMax = Inf, firstN = Inf, lastN = Inf,
  atInfection = FALSE, report = FALSE, corr = FALSE)

Arguments

epidemic
a list returned using the function simulateEpidemic in the package toyepidemic.

data
a data.table containing donor-recipient couples extracted from an epidemic simulated using the toyepidemic R -package.

sampledOnly
logical, indicating if only sampled individuals should be included in the heritability calculation

tMin, tMax
numeric, time interval for which to measure the heritability. if sampledOnly is TRUE this would be the times of sampling for the individuals; otherwise, this would be the times of infection.

corr
logical, should donor-recipient correlation be returned instead of regression slope