A box-plot of phenotypic distances against phylogenetic distances from a PP analysis

A box-plot of phenotypic distances against phylogenetic distances from a PP analysis

# S3 method for PP
boxplot(object, abs = TRUE, tauQuantileType = c("V", "D", "O",
  "qu", "Q", "M", "A"), filter = "TRUE", mapping = "aes(tau, deltaz)",
  facets = tauQuantileType ~ ., ...)

Arguments

object
An object of class PP.

abs
Logical indicating whether to plot absolute phenotypic distances.

filter
A character evaluating as an R expression. Defaults to TRUE.

mapping
A character string. Defaults to

facets
A formula passed to facet_grid 'aes(tau, deltaz)'