Plot a correlation profile

Plot a correlation profile

plotCorrProfile(object, corrProfile = corrProfile(object, tauQuantileTypes =
  tauQuantileTypes), tauQuantileTypes = c("A", "M", "Q", "D", "V"),
  palette = c("#999999", "#0072B2", "#CC79A7", "#E69F00", "#D55E00",
  "#56B4E9", "#009E73", "#F0E442"), xlim = NULL, ylim = NULL,
  showLabel = TRUE, dodgeRatio = 1/25, labelSize = 2.5,
  labelColor = "black")

Arguments

object
an S3 object of class "H2Analysis"

tauQuantileTypes
a character vector of at least 1 element denoting the types of quantiles of tau (stratifications) to plot the correlation profile for. All these types must already be available in the corrTable object.

corrTable
a data.table returned by corrProfile