Bootstrap ANOVA

Bootstrap ANOVA

rAboot(i, idPair, z, bootstraps = 1000, as.dt = FALSE, ...)

Arguments

i
integer vector identifiers of the individuals

idPair
integer vector showing the membership of individuals in pairs or bigger classes

z
numeric vector of phenotypes

bootstraps
integer the number of bootstrap samples to perform (default 1000)

...
other columns that are appended to the resulting data.table if as.dt = TRUE. This parameter is supplied for convinience only.

Value

list iwth estimated rA and 95

Details

the function performs 1000 bootstraps