Scatter plot of phylogenetic pairs

Scatter plot of phylogenetic pairs

scatterPlotPPs(z, tree, ppAnalysis, CPPthr = 10^-4, ruleOutXIQR = 1.5,
  zName = "z", treeName = "tree", xlab = expression(lg(tau) ~
  "[lg(subst. per site)]"), ylab = expression(group("|", Delta ~ lg(spVL),
  "|")), xlim = c(-6, 0), ylim = c(0, 5))

Arguments

z
numeric vector of phenotypes

tree
a phylo object with tip-labels corresponding to the entries in z

ppAnalysis
a list resulting from a call to analyseCPPs on the same tree and z.

CPPthr
numeric indicating the maximum phylogenetic distance separating a closest phylogenetic pair

ruleOutXIQR
numeric a multiplier used to define outlier CPPs as defined in the referenced article.

zName, treeName
characters used when the parameter z is a list; indicate the corresponding names in the list

xlim, ylim, xlab, ylab
graphical parameters passed to plot